requirements.txt
amqp==2.5.2
Babel==2.8.0
backcall==0.1.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==2.2.10
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
flower==0.9.3
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
ipython==7.10.2
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.2
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
zipp==2.1.0
